requirements.txt
emoji==2.8.0
flake8==6.1.0
matplotlib==3.8.2
nltk==3.8.1
numpy==1.26.2
openai==1.3.5
pandas==2.1.3
python-dotenv==1.0.0
scikit-learn==1.3.2
torch==2.1.1
transformers==4.35.2
tweepy==4.14.0
vaderSentiment==3.3.2
